+++ packages/api-contracts/src/index.ts
@@ -1,3 +1,4 @@
-export * from './userApiContracts.ts'
-export * from './userErrors.ts'
-export * from './userSchemas.ts'
+export * from './api/contracts/index.ts'
+export * from './api/errors/index.ts'
+export * from './api/objects/index.ts'
+export * from './api/open-api-tags.ts'