+++ src/modules/users/consumers/PermissionsConsumer.spec.ts
@@ -29,6 +29,7 @@ async function createUsers(drizzle: PostgresJsDatabase, userIdsToCreate: string[
         id: userId,
         name: userId.toString(),
         email: `test${userId}@email.lt`,
+        internalNote: 'internal note',
       }
     }),
   )